rec-server/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.4.0+cpu
torchvision==0.5.0+cpu
Flask
Flask-Cors
scipy
recsys-recoder==0.3.0
numpy